+++ provider/cmd/pulumi-resource-kong/main.go
@@ -17,8 +17,8 @@
 package main
 
 import (
-	kong "github.com/pulumi/pulumi-kong/provider/v3"
-	"github.com/pulumi/pulumi-kong/provider/v3/pkg/version"
+	kong "github.com/pulumi/pulumi-kong/provider/v4"
+	"github.com/pulumi/pulumi-kong/provider/v4/pkg/version"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
 )