+++ sdk/dotnet/Provider.cs
@@ -76,6 +76,12 @@ namespace Pulumi.Kong
         public Input<string>? KongApiKey { get; set; }
 
         /// <summary>
+        /// Workspace context (Enterprise Edition)
+        /// </summary>
+        [Input("kongWorkspace")]
+        public Input<string>? KongWorkspace { get; set; }
+
+        /// <summary>
         /// Should plugins `config_json` field strictly match plugin configuration
         /// </summary>
         [Input("strictPluginsMatch", json: true)]