+++ sdk/go/kong/config/config.go
@@ -33,6 +33,11 @@ func GetKongApiKey(ctx *pulumi.Context) string {
 	return config.Get(ctx, "kong:kongApiKey")
 }
 
+// Workspace context (Enterprise Edition)
+func GetKongWorkspace(ctx *pulumi.Context) string {
+	return config.Get(ctx, "kong:kongWorkspace")
+}
+
 // Should plugins `config_json` field strictly match plugin configuration
 func GetStrictPluginsMatch(ctx *pulumi.Context) bool {
 	v, err := config.TryBool(ctx, "kong:strictPluginsMatch")