+++ sdk/go/kong/provider.go
@@ -54,6 +54,8 @@ type providerArgs struct {
 	KongAdminUsername *string `pulumi:"kongAdminUsername"`
 	// API key for the kong api (if you have locked it down)
 	KongApiKey *string `pulumi:"kongApiKey"`
+	// Workspace context (Enterprise Edition)
+	KongWorkspace *string `pulumi:"kongWorkspace"`
 	// Should plugins `config_json` field strictly match plugin configuration
 	StrictPluginsMatch *bool `pulumi:"strictPluginsMatch"`
 	// Whether to skip tls verify for https kong api endpoint using self signed or untrusted certs
@@ -72,6 +74,8 @@ type ProviderArgs struct {
 	KongAdminUsername pulumi.StringPtrInput
 	// API key for the kong api (if you have locked it down)
 	KongApiKey pulumi.StringPtrInput
+	// Workspace context (Enterprise Edition)
+	KongWorkspace pulumi.StringPtrInput
 	// Should plugins `config_json` field strictly match plugin configuration
 	StrictPluginsMatch pulumi.BoolPtrInput
 	// Whether to skip tls verify for https kong api endpoint using self signed or untrusted certs